Put abstract normalizations class in abstractions.py
